requirements.txt
shortuuid
flask
Pillow